exdroid4.0.4r1-a13-v1.5release

+++ media/libstagefright/SampleIterator.cpp
@@ -306,8 +306,7 @@ status_t SampleIterator::findSampleTime(
     }
 
     *time = mTTSSampleTime + (uint64_t)mTTSDuration * (sampleIndex - mTTSSampleIndex);
-
-    *time += mTable->getCompositionTimeOffset(sampleIndex);
+    *time += (int32_t)mTable->getCompositionTimeOffset(sampleIndex);
 
     return OK;
 }
